This is still a giant mess and needs a whole lot of fixing up, but it should be mostly functional. All changes and text segments should be properly separated and mostly marked up.
Separators are still in the section files; that way, the presence of a separator
is a clear indicator that a section wasn't post-processed yet.

The source files can be converted to HTML and other formats using Sphinx (http://sphinx.pocoo.org/),
the files are written in Sphinx's variant of reStructuredText (http://sphinx.pocoo.org/rest.html).